split out the TCP part of dnsdist to a separate file
