+++ pdns/dnsdistdist/populate
@@ -4,7 +4,7 @@ ln -fs ../base32.hh ../base64.hh ../dnsdist.cc ../dnsdist.hh ../dnsdist-lua.cc .
 ../dnslabeltext.rl ../dnsname.cc ../dnsname.hh ../dnsparser.hh ../dnsrulactions.hh ../dnswriter.cc ../dnswriter.hh \
 ../dolog.hh ../iputils.cc ../iputils.hh ../misc.cc ../misc.hh ../namespaces.hh \
 ../pdnsexception.hh ../qtype.cc ../qtype.hh ../sholder.hh ../sodcrypto.cc ../sodcrypto.hh \
-../dnsdist-web.cc ../sstuff.hh .
+../dnsdist-web.cc ../sstuff.hh ../dnsdist-tcp.cc .
 
 ln -fs ../README-dnsdist.md README.md
